Data.Serialize.Descriptor

Documentation

newtype Descriptor s a

Descriptor s a is an applicative functor that describes the binary structure for a structure s while deserializing value a.

Constructors

Descriptor (Get a, s -> PutM a)

unwrapGet :: Descriptor s a -> Get a

unwrapGet desc takes a Descriptor and returns only the internal Get monad.

unwrapPut :: s -> Descriptor s a -> PutM a

unwrapPut s desc takes the structure being described and a Descriptor for it, and returns the internal PutM monad.

serialize :: s -> Descriptor s a -> ByteString

Convenience function for runPut . unwrapPut s

deserialize :: ByteString -> Descriptor s s -> Either String s

Convenience function for flip runGet bs . unwrapGet