Data.Serialize.Descriptor.LE

Description

Little endian combinators.

All combinators take a function that takes the structure being described (a) and produces the specified data type from it. Most of the time, this will be one of the structure's fields, which are all functions from the structure to the field type.

Documentation

w8 :: (a -> Word8) -> Descriptor a Word8

w16 :: (a -> Word16) -> Descriptor a Word16

w32 :: (a -> Word32) -> Descriptor a Word32

w64 :: (a -> Word64) -> Descriptor a Word64

i8 :: (a -> Int8) -> Descriptor a Int8

i16 :: (a -> Int16) -> Descriptor a Int16

i32 :: (a -> Int32) -> Descriptor a Int32

i64 :: (a -> Int64) -> Descriptor a Int64

f32 :: (a -> Float) -> Descriptor a Float

f64 :: (a -> Double) -> Descriptor a Double